--- WorQit/WorQit/Main.xaml
@@ -1,44 +1,0 @@
-﻿<Page
-    x:Class="WorQit.Main"
-    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    xmlns:local="using:WorQit"
-    xmlns:converter="using:WorQit.Converters"
-    xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-    xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-    mc:Ignorable="d">
-    <Page.Resources>
-        <converter:DateTimeConverter x:Key="Converter"/>
-        <DataTemplate x:Key="berichtTemplate">
-            <Grid Height="33" Width="440" Tapped="messageClick">
-                <Grid.ColumnDefinitions>
-                    <ColumnDefinition Width="50*"/>
-                    <ColumnDefinition Width="190*"/>
-                    <ColumnDefinition Width="150*"/>
-                </Grid.ColumnDefinitions>
-                <TextBlock x:Name="txtTitleOpen" Tapped="messageClick" TextWrapping="Wrap" Text="{Binding title}" FontSize="12" VerticalAlignment="Center" TextAlignment="Center" Grid.ColumnSpan="2" Margin="58,9,0,8" />
-                <Image x:Name="imgMail" Tapped="messageClick" Source="{Binding imgPath}" Margin="0,0,0.333,0"/>
-                <TextBlock x:Name="txtDate" Grid.Column="2" TextWrapping="Wrap" Text="{Binding Path=date, Converter={StaticResource ResourceKey=Converter}, ConverterParameter=\{dd/MM/yyyy HH:mm\}}" FontSize="12" VerticalAlignment="Center" TextAlignment="Center"/>
-            </Grid>
-        </DataTemplate>
-    </Page.Resources>
-
-    <Grid Background="{ThemeResource ApplicationPageBackgroundThemeBrush}">
-        <Grid.ColumnDefinitions>
-            <ColumnDefinition Width="97*"/>
-            <ColumnDefinition Width="168*"/>
-            <ColumnDefinition Width="96*"/>
-        </Grid.ColumnDefinitions>
-        <Grid.RowDefinitions>
-            <RowDefinition Height="29*"/>
-            <RowDefinition Height="9*"/>
-            <RowDefinition Height="112*"/>
-            <RowDefinition Height="10*"/>
-        </Grid.RowDefinitions>
-        <Image x:Name="image" Grid.Column="1" Source="Assets/logo WorQit.png" Grid.ColumnSpan="2" Margin="0.333,0,95,0"/>
-        <Button x:Name="btnBack" Content="terug" Grid.Row="3" FontSize="12" Background="#FFA3CBE8" Foreground="White" Grid.Column="1" HorizontalAlignment="Stretch" VerticalAlignment="Stretch" Margin="5.333,5,95,5" Click="btnSettings_Copy_Click" Style="{StaticResource ButtonStyle}" Grid.ColumnSpan="2"/>
-        <Rectangle Grid.ColumnSpan="3" Fill="#FF195695" Grid.Row="1" Stroke="Black"/>
-        <TextBlock Grid.Row="1" TextWrapping="Wrap" Text="berichten" Grid.Column="1" FontSize="15" Foreground="White" TextAlignment="Center" VerticalAlignment="Center"/>
-        <GridView x:Name="control" ItemClick="control_ItemClick"  Grid.Row="2" ItemsSource="{Binding berichten}" ItemTemplate="{StaticResource berichtTemplate}" Grid.ColumnSpan="3" Margin="10,0"/>
-    </Grid>
-</Page>
